advanced-rag-chatbot/apps/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
httpx==0.27.0
pydantic==2.7.1
python-dotenv==1.0.1
loguru==0.7.2
pytest==8.2.2
pytest-asyncio==0.23.6
pdfplumber==0.11.4
PyPDF2==3.0.1
sentence-transformers==3.0.1
numpy==1.26.4
scipy==1.12.0
scikit-learn==1.4.2
asyncpg==0.29.0
python-multipart==0.0.9
tenacity==8.3.0

